==> 03_Scripts/2_Importing_Libraries_and_Python_Data_Types.ipynb <==
**This script contains the following points:**

1. Importing libraries
2. Data types
    - Question 7 - Code 3 different ways of reaching a result of 100 by adding or subtracting numeric variables.
    - Question 8 - Code 2 floating-point variables and divide them by each other.
    - Question 9 - Construct a short word made of separate strings.
    - Question 10 - Construct 2 short sentences made of separate strings.
    - Bonus - Include a command that makes a string appear 100 times.

# 01. Importing libraries

# Importing libraries
import pandas as pd
import numpy as np
import os

# 02. Data types

#### *Question 7 - Code 3 different ways of reaching a result of 100 by adding or subtracting numeric variables.*

# Defining variables
x = 50
y = 30
z = 20

# First option
x + y + z

# Second option
x * 3 - x

# Third option
n = y * 4
n - z

#### *Question 8 - Code 2 floating-point variables and divide them by each other.*

x = 4.98765
y = 8.34567

y / x

#### *Question 9 - Construct a short word made of separate strings.*

# First option
'Sa' + 'r' + 'ah'

# Second option
a = 'Py'
b = 'th'
c = 'on'

a + b + c

#### *Question 10 - Construct 2 short sentences made of separate strings.*

# First sentence
a = 'Hello world,'
b = 'Sarah'
c = 'is learning Python'
space = ' '
punct = '!'

first_sentence = a + space + b + space + c + punct
print(first_sentence)

# Second sentence
e = 'She'
f = 'thinks'
g = 'it\'s'
h = 'fun'

second_sentence = e + space + f + space + g + space + h + punct
print(second_sentence)

# Making it a message
message = first_sentence + space + second_sentence
print(message)

#### *Bonus - Include a command that makes a string appear 100 times.*

# Variable b from first sentence in Question 10.
string = b  
string * 100
