==> 03_Scripts/3_Data_Import_and_Descriptive_Analysis.ipynb <==
**This script contains the following points:**

1. Importing libraries
2. Importing data
3. Data structures
    - Answers Question 11

# 01. Importing libraries

# Importing libraries (Question 3)
import pandas as pd
import numpy as np
import os

# 02. Importing data

# Turning project folder path into a string (Question 4)
path = r'/Users/sarahtischer/Desktop/CareerFoundry/Data Immersion/Achievement 4/01-2024_Instacart_Basket_Analysis'

# Defining vars_list shortcut to omit “eval_set” column (Question 5)
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

# Importing “orders.csv” dataset using the os library (Question 5)
df = pd.read_csv(os.path.join(path, '02_Data', 'Original_data', 'orders.csv'), usecols = vars_list)

# Importing “products.csv” file (Question 7)
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'Original_data', 'products.csv'), index_col = False)

# 03. Data Structures

# Printing the first 5 rows of the dataframe (Question 6)
df.head()

# Printing the first 20 rows of the df_prods dataframe (Question 8)
df_prods.head(20)

# Printing the last 35 rows of the df_prods dataframe (Question 8)
df_prods.tail(35)

# Printing names of columns in the df_prods dataframe (Question 9)
df_prods.columns

# Printing number of rows and columns in the df_prods dataframe (Question 10)
df_prods.shape

# Generating descriptive statistics of df_prods dataframe (Question 11)
df_prods.describe()

# Printing data types of df.prods dataframe (Question 11)
df_prods.dtypes

### Answers Question 11

#### *What's the maximum value of the "aisle_id" column?*

```134.000000```

#### *What's the data type of the "department_id" column?*

```int64 (Integer)```
