==> 03_Scripts/4_Data_Wrangling_and_Subsetting.ipynb <==
**This script contains the following points:**

1. Importing libraries
2. Importing data
3. Data wrangling
4. Answering questions
    - Question 2
    - Question 3
    - Question 4
    - Question 5
    - Question 6
    - Question 7
    - Question 8
    - Question 9
    - Question 10
5. Exporting data

# 01. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing data

# Turn project folder path into a string
path = r'/Users/sarahtischer/Desktop/CareerFoundry/Data Immersion/Achievement 4/01-2024_Instacart_Basket_Analysis'

# Import "orders.csv"
df_ords = pd.read_csv(os.path.join(path, '02_Data', 'Original_data', 'orders.csv'), index_col = False)

df_ords.shape

# Import "products.csv"
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'Original_data', 'products.csv'), index_col = False)

df_prods.shape

# Import "departments.csv"
df_dep = pd.read_csv(os.path.join(path, '02_Data', 'Original_data', 'departments.csv'), index_col = False)

df_dep.shape

# 03. Data wrangling

# Drop 'eval_set' column from df_ords
df_ords = df_ords.drop(columns = ['eval_set'])

# Rename 'order_dow' to 'orders_day_of_week' in df_ords
df_ords.rename(columns = {'order_dow' : 'orders_day_of_week'}, inplace = True)

# Change data type of 'order_id' from int to str
df_ords['order_id'] = df_ords['order_id'].astype('str')

# Transpose df_dep to df_dep_t
df_dep_t = df_dep.T

# Add an index to df_dep_t
df_dep_t.reset_index()

# Assign a new header to df_dep_t

new_header = df_dep_t.iloc[0] # Assigning values of first row (index = 0) to a variable

df_dep_t_new = df_dep_t[1:] # Creating new dataframe by copying rows beyond the first row

df_dep_t_new.columns = new_header # Assigning new_header variable as column names

df_dep_t_new

# Create a data dictionary of df_dep_t_new
data_dict = df_dep_t_new.to_dict('index')

data_dict

# Create a subset of df_prods
df_snacks = df_prods[df_prods['department_id']==19]

# 04. Answering questions

df_ords.info()

### Question 2

# Change data type of 'user_id' from int to str
df_ords['user_id'] = df_ords['user_id'].astype('str')

### Question 3

# Rename 'orders_day_of_week' to 'order_day_of_week' in df_ords to be consistent with 'order_hour_of_day',
# Rename 'order_number' to 'order_sequence' in df_ords to make it more intuitive
df_ords.rename(
    columns = {'orders_day_of_week' : 'order_day_of_week', 'order_number' : 'order_sequence'}, 
    inplace = False
)

### Question 4

# Find frequencies of order hours
df_ords['order_hour_of_day'].value_counts(dropna = True)

#### *<mark>Answer:</mark> The busiest hour for placing orders is 10 a.m.*

### Question 5

# Determine the meaning behind a value of 4 in the 'department_id' column
print(data_dict.get('4'))

#### *<mark>Answer:</mark> Products with a department ID of 4 are in the department 'produce'.*

### Question 6

# Create subset with breakfast item sales
df_breakfast = df_prods.loc[df_prods['department_id'] == 14]

### Question 7

# Create subset for dinner party items from departments alcohol, deli, beverages, and meat/seafood
df_dinner_party = df_prods.loc[df_prods['department_id'].isin([5, 7, 12, 20])]

df_dinner_party

### Question 8

# Print dimensions of df_dinner_party
df_dinner_party.shape

#### *<mark>Answer:</mark> df_dinner_party has 7650 rows.*

### Question 9

# Print information for user_id = 1 as df_ords_user_1
df_ords_user_1 = df_ords.loc[df_ords['user_id'] == '1']

### Question 10

# Print descriptive statistics for df_ords_user_1
df_ords_user_1.describe()

#### *<mark>Answers</mark>*
* ***The user has placed 11 orders in total.***
* ***The user ordered earliest at 7 a.m. and latest at 4 p.m., and usually orders around 10 a.m.***
* ***The user placed orders between day 1 (Sunday) and 4 (Wednesday) of the week.***
* ***On average, the user orders every 19 days.***

# 05. Exporting data

df_ords.shape

# Export df_ords as "orders_wrangled.csv"
df_ords.to_csv(os.path.join(path, '02_Data','Prepared_data', 'orders_wrangled.csv'), index = False)

df_dep_t_new.shape

# Export df_dep_t_new as "departments_wrangled.csv"
df_dep_t_new.to_csv(os.path.join(path, '02_Data','Prepared_data', 'departments_wrangled.csv'))
