==> 03_Scripts/5_Data_Consistency_Checks.ipynb <==
**This script contains the following points:**

    1. Importing libraries
    2. Importing data
    3. Consistency checks
        3.1 products dataframe
            3.1.1 Missing values
            3.1.2 Mixed-type data
            3.1.3 Duplicates
        3.2 orders_wrangled dataframe
            3.2.1 Missing values
            3.2.2 Mixed-type data
            3.2.3 Duplicates   
    4. Exporting data

# 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 2. Importing data

# Turn project folder path into a string
path = r'/Users/sarahtischer/Desktop/CareerFoundry/Data Immersion/Achievement 4/01-2024_Instacart_Basket_Analysis'

# Import "products.csv"
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'Original_data', 'products.csv'), index_col = False)

df_prods.shape

# Import "orders_wrangled.csv"
df_ords = pd.read_csv(os.path.join(path, '02_Data', 'Prepared_data', 'orders_wrangled.csv'), index_col = False)

df_ords.shape

# 3. Consistency checks

## 3.1 products dataframe

# Suppress scientific notation by specifying the display format for floating-point numbers
pd.options.display.float_format = '{:,.2f}'.format

# Print descriptive statistics (Question 2)
df_prods.describe()

#### *<mark>Answer:</mark> The descriptive statistics for the products dataframe show a maximum value for the 'prices' column of 99,999.00, which seems unusually high for grocery prices. It is possible that there is an outlier or an error in the data. Realistic prices are typically not that high.*

### 3.1.1 Missing values

# Find missing values
df_prods.isnull().sum() # Find missing observations and sum results by column

# Create subset of missing observations
df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

# Print dimensions of df_prods
df_prods.shape

# Create subset without missing observations
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# Print dimensions of df_prods_clean
df_prods_clean.shape

### 3.1.2 Mixed-type data

# Check for mixed types (Question 3)
for col in df_prods_clean.columns.tolist():
  mixed = (df_prods_clean[[col]].map(type) != df_prods_clean[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_prods_clean[mixed]) > 0:
    print (col)

### 3.1.3 Duplicates

# Find full duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

# Print dimensions of df_prods_clean
df_prods_clean.shape

# Create subset without duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

# Print dimensions of df_prods_clean_no_dups
df_prods_clean_no_dups.shape

## 3.2 orders_wrangled dataframe

# Suppress scientific notation by specifying the display format for floating-point numbers
pd.options.display.float_format = '{:,.2f}'.format

# Print descriptive statistics (Question 2)
df_ords.describe()

#### *<mark>Answer:</mark> The descriptive statistics for the orders dataframe don't show any unreasonable values.*

### 3.2.1 Missing values

# Find missing values (Question 5)
df_ords.isnull().sum()

# Create subset of missing observations
df_empty = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_empty

#### *<mark>Answer:</mark> The instances of NaN values in the days_since_prior_order column are not really missing values. In fact, they correlate with order_number = 1, which means that this is the user's first order, and therefore there has been no prior order and no time since. Hence, these NaN values should be kept.*

#### *<mark>Note:</mark> No action required for Question 6.*

### 3.2.2 Mixed-type data

# Check for mixed types (Question 3)
for col in df_ords.columns.tolist():
  nonconforming = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[nonconforming]) > 0:
    print (col)

#### *<mark>Answer:</mark> The orders dataframe doesn't contain any mixed-type columns.*

#### *<mark>Note:</mark> No action required for Question 4.*

### 3.2.3 Duplicates

# Find full duplicates (Question 7)
df_dups_ords = df_ords[df_ords.duplicated()]

df_dups_ords

#### *<mark>Answer:</mark> There are no duplicates in the orders dataframe.*

#### *<mark>Note:</mark>No action required for Question 8.*

# 04. Exporting data

df_prods_clean_no_dups.shape

# Export df_prods_clean_no_dups as "products_checked.csv"
df_prods_clean_no_dups.to_csv(os.path.join(path, '02_Data','Prepared_data', 'products_clean.csv'), index = False)

df_ords.shape

# Export df_ords as "orders_checked.csv"
df_ords.to_csv(os.path.join(path, '02_Data','Prepared_data', 'orders_clean.csv'), index = False)
