==> 03_Scripts/6_Combining_and_Exporting_Data_(E).ipynb <==
**This script contains the following points:**

1. Importing libraries
2. Importing data
3. Merging data
4. Exporting data

# 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 2. Importing data

# Turn project folder path into a string
path = r'/Users/sarahtischer/Desktop/CareerFoundry/Data Immersion/Achievement 4/01-2024_Instacart_Basket_Analysis'

# Import "orders_products_prior.csv"
df_ords_prior = pd.read_csv(os.path.join(path, '02_Data', 'Original_data', 'orders_products_prior.csv'), index_col = False)

# Import "orders_clean.csv"
df_ords = pd.read_csv(os.path.join(path, '02_Data', 'Prepared_data', 'orders_clean.csv'), index_col = False)

# 3. Merging data

# Check the output of df_ords_prior
df_ords_prior.head()

df_ords_prior.shape

# Check the output of df_ords
df_ords.head()

df_ords.shape

# Merge df_ords and df_ords_prior
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

# Check the output of df_merged_large
df_merged_large.head()

df_merged_large['_merge'].value_counts()

df_merged_large.shape

# 4. Exporting data

# Export df_merged_large as "orders_products_combined.pkl"
df_merged_large.to_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'orders_products_combined.pkl'))
