==> 03_Scripts/6_Combining_and_Exporting_Data_(T).ipynb <==
**This script contains the following points:**

1. Importing libraries
2. Importing data
3. Merging data
4. Exporting data

# 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 2. Importing data

# Turn project folder path into a string
path = r'/Users/sarahtischer/Desktop/CareerFoundry/Data Immersion/Achievement 4/01-2024_Instacart_Basket_Analysis'

# Import "products_clean.csv"
df_prods = pd.read_csv(os.path.join(path, '02_Data', 'Prepared_data', 'products_clean.csv'), index_col = False)

# Check the output of df_prods
df_prods.head()

df_prods.shape

# Import "orders_products_combined.pkl" (Question 3)
df_ords_prods_comb = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'orders_products_combined.pkl'))

# Check the output of df_ords_prods_comb (Question 4)
df_ords_prods_comb.head()

df_ords_prods_comb.shape

# 3. Merging data

# Drop _merge column from previous merge in df_ords_prods
df_ords_prods = df_ords_prods_comb.drop(columns = ['_merge'])

# Check the output of df_ords_prods
df_ords_prods.head()

df_ords_prods.shape

# Test full join
#pd.merge(df_prods, df_ords_prods, on = ['product_id'], indicator = True, how = 'outer')

# Shape output:
# 32435070 rows × 14 columns

# Merge df_prods and df_ords_prods using 'inner' join as a full merge set is required (Question 5)
df_ords_prods_merged = df_prods.merge(df_ords_prods, on = ['product_id'], indicator = True)

# Check the output of df_ords_prods_merge
df_ords_prods_merged['_merge'].value_counts()

df_ords_prods_merged.shape

# 4. Exporting data

# Export df_ords_prods_merge as "orders_products_merged.pkl" 
df_ords_prods_merged.to_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'orders_products_merged.pkl'))
