==> 03_Scripts/7_Deriving_New_Variables.ipynb <==
**This script contains the following points:**

1. Importing libraries
2. Importing data
3. Deriving new variables
    - If-statements with user-defined functions
    - If-statements with loc() functions
        - Dataframe: df_ords_prods
        - Dataframe: df_ords_prods_merge
    - If-statements with for-loops
4. Answering business questions
    - 'busiest_days'
    - 'busiest_period_of_day'
5. Exporting data

# 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 2. Importing data

# Turn project folder path into a string
path = r'/Users/sarahtischer/Desktop/CareerFoundry/Data Immersion/Achievement 4/01-2024_Instacart_Basket_Analysis'

# Import "orders_products_merged.pkl"
df_ords_prods_merged = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'orders_products_merged.pkl'))

df_ords_prods_merged.shape

# Limit the dataframe to one million rows
df_sub = df_ords_prods_merged[:1000000]

df_sub.shape

# 3. Deriving new variables

## If-statements with user-defined functions

# Define function price_label
def price_label(row):
    if row['prices'] <= 5:
        return 'Low-range product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'Mid-range product'
    elif row['prices'] > 15:
        return 'High-range product'
    else:
        return np.nan

# Apply function price_label
df_sub['price_range'] = df_sub.apply(price_label, axis = 1)

# Check frequencies of new row price_range
df_sub['price_range'].value_counts(dropna = False)

df_sub['prices'].max()

## If-statements with loc() function

### Dataframe: df_ords_prods

# Create conditions for price_range_loc
df_sub.loc[df_sub['prices'] <= 5, 'price_range_loc'] = 'Low-range product'
df_sub.loc[(df_sub['prices'] > 5) & (df_sub['prices'] <= 15), 'price_range_loc'] = 'Mid-range product'
df_sub.loc[df_sub['prices'] > 15, 'price_range_loc'] = 'High-range product'

# Check frequencies of new column 'price_range_loc'
df_sub['price_range_loc'].value_counts(dropna = False)

### Dataframe: df_ords_prods_merged

# Create conditions for price_range_loc
df_ords_prods_merged.loc[df_ords_prods_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'
df_ords_prods_merged.loc[(df_ords_prods_merged['prices'] <= 15) & (df_ords_prods_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'
df_ords_prods_merged.loc[df_ords_prods_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'

# Check frequencies of new column 'price_range_loc'
df_ords_prods_merged['price_range_loc'].value_counts(dropna = False)

## If-statements with for-loops

# Determine the busiest day
df_ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

# Create result list
result = []

# Run for-loop
for value in df_ords_prods_merged['orders_day_of_week']:
  if value == 0:
    result.append('Busiest day')
  elif value == 4:
    result.append('Least busy')
  else:
    result.append('Regularly busy')

# Print result list
result

# Assign result list to df_ords_prods_merged column 'busiest_day'
df_ords_prods_merged['busiest_day'] = result

# Check frequencies of new column 'busiest_day'
df_ords_prods_merged['busiest_day'].value_counts(dropna = False)

# 4. Answering business questions

## *'busiest_days'*

# Rename 'busiest_day' to 'busiest_days'
df_ords_prods_merged.rename(columns = {'busiest_day' : 'busiest_days'}, inplace = True)

# Determine the busiest and slowest days
df_ords_prods_merged['orders_day_of_week'].value_counts(dropna = False)

# Create result list
result_2 = []

# Run for-loop (Question 2)
for value in df_ords_prods_merged['orders_day_of_week']:
  if value in (0, 1):
    result_2.append('Busiest day')
  elif value in (3, 4):
    result_2.append('Least busy')
  else:
    result_2.append('Regularly busy')

# Print result list
result_2

# Assign result list to df_ords_prods_merged column 'busiest_days'
df_ords_prods_merged['busiest_days'] = result_2

# Check values of new column for accuracy (Question 3)
df_ords_prods_merged.head(15)

#### *<mark>Answer:</mark> All values (0 - 6) have been assigned the correct string label in the 'busiest_days' column.*

# Check frequencies of new column 'busiest_days'
df_ords_prods_merged['busiest_days'].value_counts(dropna = False)

## *'busiest_period_of_day'*

# Determine the busiest hours of the day
df_ords_prods_merged['order_hour_of_day'].value_counts(dropna = False)

# Create result list
result_3 = []

# Create *_orders tuples for easier access
most_orders = (10, 11, 14, 15, 13, 12, 16, 9)
fewest_orders = (23, 6, 0, 1, 5, 2, 4, 3)

# Run for-loop (Question 4)
for value in df_ords_prods_merged['order_hour_of_day']:
  if value in most_orders:
    result_3.append('Most orders')
  elif value in fewest_orders:
    result_3.append('Fewest orders')
  else:
    result_3.append('Average orders')

# Print result list
result_3

# Assign result list to df_ords_prods_merged column 'busiest_period_of_day' (Question 4)
df_ords_prods_merged['busiest_period_of_day'] = result_3

# Check frequencies of new column 'busiest_period_of_day' (Question 5)
df_ords_prods_merged['busiest_period_of_day'].value_counts(dropna = False)

# 5. Exporting data

df_ords_prods_merged.shape

# Export df_ords_prods_merged as "orders_products_merged(2).pkl" 
df_ords_prods_merged.to_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'orders_products_merged(2).pkl'))
