==> 03_Scripts/9_Data_Visualization_with_Python_(1).ipynb <==
**This script contains the following points:**

    1. Importing libraries
    2. Importing data
    3. Wrangling data (Question 4)
        3.1 Renaming columns
        3.2 Changing data types  
    4. Data quality and consistency checks (Question 5)
        4.1 Missing values
        4.2 Mixed-type values
        4.3 Duplicates
    5. Combining data
        5.1 Preparing df_ords_prods for merge
            Check general structure and metadata
            Change data types and drop redundant columns
            Locate outliers and mark them NaN
        5.2 Merging dataframes (Question 6)
    6. Exporting data

# 1. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 2. Importing data

# Turn project folder path into a string
path = r'/Users/sarahtischer/Desktop/CareerFoundry/Data Immersion/Achievement 4/01-2024_Instacart_Basket_Analysis'

# Import "customers.csv"
df_cust = pd.read_csv(os.path.join(path, '02_Data', 'Original_data', 'customers.csv'))

df_cust.shape

# Import "orders_products_merged(3).pkl"
df_ords_prods = pd.read_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'orders_products_merged(3).pkl'))

df_ords_prods.shape

# 3. Wrangling data

# Check structure of "customers.csv" dataframe
df_cust.head()

df_cust.tail()

## 3.1 Renaming columns

# Rename columns
df_cust.rename(
    columns = {'First Name' : 'first_name', 
               'Surnam' : 'last_name',
               'Gender' : 'gender',
               'STATE' : 'state',
               'Age' : 'age',
               'n_dependants' : 'dependants',
               'fam_status' : 'family_status'
              }, 
    inplace = True
)

# Check the output
df_cust.head()

## 3.2 Changing data types

# Print info of df_cust
df_cust.info()

# Change data types
df_cust['user_id'] = df_cust['user_id'].astype('str')
df_cust['age'] = df_cust['age'].astype('int8')
df_cust['dependants'] = df_cust['dependants'].astype('int8')
df_cust['income'] = df_cust['income'].astype('int32')

# Check the output
df_cust.dtypes

# 4. Data quality and consistency checks

# Suppress scientific notation by specifying the display format for floating-point numbers
pd.options.display.float_format = '{:,.2f}'.format

# Print descriptive statistics
df_cust.describe()

#### *<mark>Notes</mark>*

1. Income Distribution:

* The minimum income is \$25,903, which seems reasonable.
* The maximum income is \$593,901, which is significantly higher than the 75th percentile (Q3). This could indicate potential outliers or extreme values.

2. Age Distribution:

* The minimum age is 18, which is typical for adult users.
* The maximum age is 81, and the mean age is 49.5. The maximum age doesn't seem unusual by itself, but it's essential to consider the context of the dataset and the nature of the Instacart platform.

To further investigate potential anomalies:

* Examine the distribution of income visually, considering a histogram or boxplot.
* Check for extreme values or outliers in both age and income.

## 4.1 Missing values

# Check for missing values
df_cust.isnull().sum()

# Assign rows with missing values to df_nan and print the output
df_nan = df_cust[df_cust['first_name'].isnull() == True]

df_nan

#### *<mark>Note:</mark> The customers dataframe contains 11259 NaN values in the column 'first_name'. However, since names are PII and shouldn't be used in the analysis, it's fine to disregard the missing values.*

## 4.2 Mixed-type values

# Check for mixed types
for col in df_cust.columns.tolist():
  mixed = (df_cust[[col]].map(type) != df_cust[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_cust[mixed]) > 0:
    print (col)

# Unify data type of 'df_cust' to str
df_cust['first_name'] = df_cust['first_name'].astype('str')

# Check the output
df_cust.dtypes

## 4.3 Duplicates

# Find full duplicates
df_dups = df_cust[df_cust.duplicated()]

df_dups

#### *<mark>Note:</mark> The customers dataframe doesn't contain any duplicates.*

# 5. Combining data

## 5.1 Preparing df_ords_prods for merge

#### Check general structure and metadata

df_ords_prods.head()

df_ords_prods.info()

df_ords_prods.describe()

#### Change data types, rename columns and drop redundant columns

# Rename columns
df_ords_prods.rename(
    columns = {'orders_day_of_week' : 'order_day_of_week', 
               'price_range_loc' : 'price_range',
              }, 
    inplace = True
)

# Change data types to reduce size
df_ords_prods['product_id'] = df_ords_prods['product_id'].astype('int32')
df_ords_prods['aisle_id'] = df_ords_prods['aisle_id'].astype('int16')
df_ords_prods['department_id'] = df_ords_prods['department_id'].astype('int8')
df_ords_prods['prices'] = df_ords_prods['prices'].astype('float32')
df_ords_prods['order_id'] = df_ords_prods['order_id'].astype('int32')
df_ords_prods['user_id'] = df_ords_prods['user_id'].astype('str')
df_ords_prods['order_number'] = df_ords_prods['order_number'].astype('int8')
df_ords_prods['order_day_of_week'] = df_ords_prods['order_day_of_week'].astype('int8')
df_ords_prods['order_hour_of_day'] = df_ords_prods['order_hour_of_day'].astype('int8')
df_ords_prods['days_since_prior_order'] = df_ords_prods['days_since_prior_order'].astype('float32')
df_ords_prods['add_to_cart_order'] = df_ords_prods['add_to_cart_order'].astype('int16')
df_ords_prods['reordered'] = df_ords_prods['reordered'].astype('int8')
df_ords_prods['max_order'] = df_ords_prods['max_order'].astype('int8')
df_ords_prods['avg_price'] = df_ords_prods['avg_price'].astype('float16')
df_ords_prods['median_frequency'] = df_ords_prods['median_frequency'].astype('float16')

# Drop redundant variables
df_ords_prods = df_ords_prods.drop(columns = ['aisle_id', '_merge'])

# Check the output
df_ords_prods.info()

#### Locate outliers and mark them NaN

# Locate outliers in 'prices' column & mark them NaN
df_ords_prods.loc[df_ords_prods['prices'] > 100, 'prices'] = np.nan

# Check the output
df_ords_prods['prices'].max()

## 5.2 Merging dataframes

# Merge df_ords_prods and df_cust using 'inner' join as a full merge set is required
df_all_merged = df_ords_prods.merge(df_cust, on = ['user_id'], indicator = True)

# Check the output
df_all_merged['_merge'].value_counts()

df_all_merged.head()

df_all_merged.shape

# 6. Exporting data

%%time
# Export df_all_merged as "orders_products_customers_merged.pkl" 
df_all_merged.to_pickle(os.path.join(path, '02_Data', 'Prepared_data', 'orders_products_customers_merged.pkl'))
